Conflicts:
	src/com/androidodc/eorder/datatypes/Category.java
	src/com/androidodc/eorder/datatypes/Config.java
	src/com/androidodc/eorder/datatypes/DiningTable.java
	src/com/androidodc/eorder/datatypes/Dish.java
	src/com/androidodc/eorder/datatypes/DishCategory.java
	src/com/androidodc/eorder/datatypes/Order.java
	src/com/androidodc/eorder/utils/LogUtils.java

+++ src/com/androidodc/eorder/database/DatabaseHelper.java
@@ -98,7 +98,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         return CategoryTable.getAllCategorys(getReadableDatabase());
     }
 
-    public List<Dish> getDishsByCategory(int categoryId) {
+    public List<Dish> getDishsByCategory(long categoryId) {
         return DishCategoryTable.getDishsByCategory(getReadableDatabase(), categoryId);
     }
