+++ src/com/androidodc/eorder/datatypes/Category.java
@@ -5,25 +5,25 @@ import java.io.Serializable;
 public class Category implements Serializable {
 
     private static final long serialVersionUID = 3948922015270201885L;
-    private int mId;
-    private int mCategoryId;
+    private long mId;
+    private long mCategoryId;
     private String mName;
     private String mDescription;
     private int mSortOrder;
 
-    public int getId() {
+    public long getId() {
         return mId;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         mId = id;
     }
 
-    public int getCategoryId() {
+    public long getCategoryId() {
         return mCategoryId;
     }
 
-    public void setCategoryId(int categoryId) {
+    public void setCategoryId(long categoryId) {
         mCategoryId = categoryId;
     }
